refactor: 코드 리팩토링

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -2,6 +2,7 @@ package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
@@ -14,9 +15,11 @@ import java.util.stream.Collectors;
 @Transactional
 public class StationService {
     private StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     @Transactional
@@ -45,6 +48,11 @@ public class StationService {
         if (stationDao.notExistId(id)) {
             throw new NotFoundStationException();
         }
+
+        if (sectionDao.existSectionByStationId(id)) {
+            throw new IncludedStationException();
+        }
+
         stationDao.deleteById(id);
     }
 }
